+++ db/updates/update_2.0.5-2.0.6.sql
@@ -19,3 +19,14 @@ ALTER TABLE `ingredients` ADD `cat11A` varchar(255) DEFAULT NULL;
 ALTER TABLE `ingredients` ADD `cat11B` varchar(255) DEFAULT NULL;
 ALTER TABLE `ingredients` ADD `cat12` varchar(255) DEFAULT NULL;
 ALTER TABLE `settings` ADD `pubChem` INT(11) DEFAULT NULL;
+
+CREATE TABLE `makeFormula` (
+ `id` int(11) NOT NULL AUTO_INCREMENT,
+ `fid` varchar(255) COLLATE utf8_bin DEFAULT NULL,
+ `name` varchar(255) COLLATE utf8_bin NOT NULL,
+ `ingredient` varchar(255) COLLATE utf8_bin DEFAULT NULL,
+ `concentration` decimal(5,2) DEFAULT 100.00,
+ `dilutant` varchar(255) COLLATE utf8_bin DEFAULT NULL,
+ `quantity` decimal(8,2) DEFAULT NULL,
+ PRIMARY KEY (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin